refactor: Moving DynamoDB into it's own package
