+++ backend-api/src/models/index.ts
@@ -1,1 +1,1 @@
-export { User } from './user';
+export * from './user';
